requirements.txt
Django==3.1.5
python-telegram-bot==13.1
requests==2.25.1
python-decouple==3.3